auto zoom for display and new primes

+++ prime.h
@@ -11,7 +11,7 @@ struct prime
  
 typedef prime* llist;
 
-void draw_point_on_circle(SDL_Renderer* renderer, SDL_Point center, double radian, int radius, SDL_Color color);
+void draw_point_on_circle(SDL_Renderer *renderer, SDL_Point center, double radian, double radius, SDL_Color color);
 int draw_n_primes(SDL_Renderer *renderer, SDL_Point center, int window_length, int expected_number_of_primes);
 void draw_logarithm(SDL_Renderer* renderer,  SDL_Point center, double a, int expected_number_of_primes);
 void draw_big_point_on_circle(SDL_Renderer *renderer, SDL_Point center, double radian, int radius, SDL_Color color);
